Split generate.rs into submodules.
